+++ src/app/[locale]/(client)/[...rest]/page.tsx
@@ -1,0 +1,3 @@
+export default function InnerNotFoundPage() {
+  return <p>NotFound Page</p>;
+}
