+++ src/app/[locale]/(client)/not-found.tsx
@@ -1,0 +1,5 @@
+import { useTranslations } from "next-intl";
+
+export default function NotFoundPage() {
+  useTranslations("NotFoundPage");
+}
