+++ src/app/[locale]/(client)/page.tsx
@@ -1,5 +1,5 @@
 import { useTranslations } from "next-intl";
-import WelcomeBlock from "./../components/WelcomeBlock/WelcomeBlock";
+import WelcomeBlock from "../../../components/WelcomeBlock/WelcomeBlock";
 
 export default function Home() {
   const t = useTranslations("Index");
