+++ src/app/[locale]/layout.tsx
@@ -1,44 +1,9 @@
-import Footer from "./../components/Footer/Footer";
-import Header from "./../components/Header/Header";
-import "../globals.scss";
-import type { Metadata } from "next";
-import { notFound } from "next/navigation";
-import { NextIntlClientProvider } from "next-intl";
+import { ReactNode } from "react";
 
-export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+type Props = {
+  children: ReactNode;
 };
 
-async function getMessages(locale: string) {
-  try {
-    return (await import(`../../../messages/${locale}.json`)).default;
-  } catch (error) {
-    notFound();
-  }
-}
-
-export async function generateStaticParams() {
-  return ["uk", "en"].map((locale) => ({ locale }));
-}
-
-export default async function LocaleLayout({
-  children,
-  params: { locale },
-}: {
-  children: React.ReactNode;
-  params: { locale: string };
-}) {
-  const messages = await getMessages(locale);
-  return (
-    <html lang={locale}>
-      <body>
-        <NextIntlClientProvider locale={locale} messages={messages}>
-          <Header />
-          {children}
-          <Footer />
-        </NextIntlClientProvider>
-      </body>
-    </html>
-  );
+export default function RootLayout({ children }: Props) {
+  return children;
 }
