--- src/app/layout.tsx
@@ -1,9 +1,0 @@
-import { ReactNode } from "react";
-
-type Props = {
-  children: ReactNode;
-};
-
-export default function RootLayout({ children }: Props) {
-  return children;
-}
