--- src/app/not-found.tsx
@@ -1,9 +1,0 @@
-export default function NotFoundPage() {
-  return (
-    <html lang="en">
-      <body>
-        <p>NotFound Page</p>
-      </body>
-    </html>
-  );
-}
