--- src/app/page.tsx
@@ -1,5 +1,0 @@
-import { redirect } from "next/navigation";
-
-export default function RootPage() {
-  redirect("/ua");
-}
