+++ src/components/Footer/Footer.module.scss
@@ -1,6 +1,6 @@
 // @import "../../styles/globals.scss";
-@use "../../../styles//mixins" as mixins;
-@import "../../../styles/variables.scss";
+@use "./../../styles//mixins" as mixins;
+@import "./../../styles/variables.scss";
 
 .footer {
   margin-top: 50px;
