+++ src/components/WelcomeBlock/WelcomeBlock.module.scss
@@ -1,5 +1,5 @@
-@use "../../../styles/mixins" as mixins;
-@use "../../../styles/variables";
+@use "./../../styles/mixins" as mixins;
+@use "./../../styles/variables";
 
 .blockWrapper,
 .imageWrapper {
