+++ src/components/ui/Button/Button.module.scss
@@ -1,5 +1,5 @@
-@use "../../../../styles/mixins" as mixins;
-@use "../../../../styles/variables";
+@use "./../../../styles/mixins" as mixins;
+@use "./../../../styles/variables";
 
 .button {
   width: 100%;
