+++ src/i18n.ts
@@ -7,6 +7,6 @@ export default getRequestConfig(async ({ locale }) => {
   if (!locales.includes(locale as any)) notFound();
 
   return {
-    messages: (await import(`../messages/${locale}.json`)).default,
+    messages: (await import(`../localization/${locale}.json`)).default,
   };
 });
